+++ src/components/SearchButton.vue
@@ -3,19 +3,13 @@
 
 <template>
     <button @click.prevent="search()">
-        <slot>{{ label }}</slot>
+        <slot>Search</slot>
     </button>
 </template>
 
 <script>
     export default {
         inject: ['searchStore'],
-        props: {
-            label: {
-                type: String,
-                default: 'Search',
-            },
-        },
         methods: {
             search() {
                 return this.searchStore.refresh()